[WIP] First cruds

+++ back/app/schemas/product_purchase.py
@@ -5,4 +5,4 @@ class ProductPurchase(BaseModel):
     purchase_id: int
     product_id: int
     quantity: int
-    unit_price: int
+    unit_price: float
