+++ back/app/schemas/product_sale.py
@@ -5,4 +5,4 @@ class ProductSale(BaseModel):
     product_id: int
     sale_id: int
     quantity: int
-    unit_price: int
+    unit_price: float
